12:55 25/04 -> Dani

+++ Personaje.h
@@ -18,8 +18,8 @@ public:
     //Se emplea virtual para que, si se crea una instancia de una clase hija de personaje, se llame a los métodos de dicha clase y no a los genericos
     void render(sf::RenderWindow &window);
     
-    sf::Vector2<float> getPosAnterior();
-    sf::Vector2<float> getPosActual();
+    sf::Vector2<int> getPosAnterior();
+    sf::Vector2<int> getPosActual();
     sf::Sprite* getSprite();
     sf::Texture* getTexture();
     int getVida();
@@ -31,8 +31,8 @@ public:
     sf::FloatRect* getBoundingBox();
     //sf::FloatRect getBoundingBox();
     
-    void setPosAnterior(sf::Vector2<float> v);
-    void setPosActual(sf::Vector2<float> v);
+    void setPosAnterior(sf::Vector2<int> v);
+    void setPosActual(sf::Vector2<int> v);
     void setSprite(sf::Sprite* s);
     void setTexture(sf::Texture* t);
     void setVida(int v);
@@ -53,8 +53,8 @@ protected:
     bool muriendo;//Booleano para determinar si el personaje esta en animacion de muerte
     sf::FloatRect* boundingBox;
     
-    sf::Vector2<float> posAnterior;
-    sf::Vector2<float> posActual;
+    sf::Vector2<int> posAnterior;
+    sf::Vector2<int> posActual;
 };
 
 #endif	/* PERSONAJE_H */
